f1data pandas
Using pandas and f1 csv files created a simple simulation game. The user inputs amount of simulation races it wants to complete. Then pairs of racers will be randomized and the user will have to guess who will win. At random the winner will be chosen. After all the simulations are ran the total number of wins and loses will be calculated.