"Text reminder for town"... Project 1
Project 1 of a python course, I created a text reminder for any group of people. This specifically was meant for a town to remind homeowners when to take out the trash and when alternate side parking is. This use twilio for the texting function but there may be other free ways to send texts through the internet.