Sign Language Program (Final Project)
This is an application that is meant to teach/show American Sign Language. It is a basic application that has the user learn sign language at their own pace with their own choice of words and numbers they would like to learn. It is meant to have people be able to converse in sign with others.